+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2026.03.19
 - Add trigger Swimming
 - Add trigger Combat - mostly working
+- Removed trigger fast travel
 
 ## 2026.03.18
 - Added trigger Fast Travel